update return type

+++ inc/kv.h
@@ -18,6 +18,6 @@ kv_t * kv_init(size_t capacity);
 
 void kv_free(kv_t * table);
 
-int kv_put(kv_t * table, char * key, char * value);
+size_t kv_put(kv_t * table, char * key, char * value);
 
 #endif //KV_H
